+++ app/service/test/Db3.php
@@ -10,8 +10,6 @@ use Flex\Banana\Utils\Requested;
 use Flex\Banana\Classes\Db\DbManager;
 use Flex\Banana\Classes\Db\DbCipherGeneric;
 use Flex\Banana\Adapters\DbAdapter;
-use Flex\Banana\Classes\Db\WhereHelper;
-use Flex\Banana\Classes\Db\WhereSql;
 use Flex\Banana\Classes\Paging\Relation;
 use Flex\Banana\Classes\Request\FormValidation as Validation;
 use Flex\Banana\Classes\Cipher\PasswordHash;
@@ -29,7 +27,7 @@ class Db3 extends DbAdapter implements ListInterface,InsertInterface,EditUpdateI
         DbManager $db,
         DbCipherGeneric $dbCipher
     ) {
-        parent::__construct(db: $db, whereHelper: new WhereHelper(new WhereSql()));
+        parent::__construct(db: $db);
 
         # cipher
         $this->dbCipher = $dbCipher;
@@ -68,12 +66,12 @@ class Db3 extends DbAdapter implements ListInterface,InsertInterface,EditUpdateI
         # query
         $rlt = $this->db->table(R::tables('users'))
             ->select(
-                "id", 
+                "_id", 
                 $this->dbCipher->decrypt("username")." as username",
                 $this->dbCipher->decrypt("email")." as email",
                 "passwd"
             )
-            ->orderBy('id DESC')
+            ->orderBy('_id DESC')
             ->limit($paging->qLimitStart, $paging->qLimitEnd)
             ->query();
         while ( $row = $rlt->fetch_assoc() )
@@ -153,13 +151,13 @@ class Db3 extends DbAdapter implements ListInterface,InsertInterface,EditUpdateI
         # check data db
         $model->data = $this->db->table( R::tables('users'))
             ->select(
-                "id", 
+                "_id", 
                 $this->dbCipher->decrypt("username")." as username",
                 $this->dbCipher->decrypt("email")." as email"
             )
-            ->where("id", $this->requested->id)
+            ->where("_id", $this->requested->id)
             ->query()->fetch_assoc();
-        if(!isset($model->data["id"])){
+        if(!isset($model->data["_id"])){
             return  JsonEncoder::toJson(["result"=>"false","msg_code"=>"e_db_unenabled", "msg"=>R::sysmsg('e_db_unenabled')]);
         }
 
@@ -192,9 +190,9 @@ class Db3 extends DbAdapter implements ListInterface,InsertInterface,EditUpdateI
 
         # check data db
         $model->data = $this->db->table( R::tables('users'))
-            ->where("id", $this->requested->id)
+            ->where("_id", $this->requested->id)
             ->query()->fetch_assoc();
-        if(!isset($model->data["id"])){
+        if(!isset($model->data["_id"])){
             return  JsonEncoder::toJson(["result"=>"false","msg_code"=>"e_db_unenabled", "msg"=>R::sysmsg('e_db_unenabled')]);
         }
 
@@ -204,7 +202,7 @@ class Db3 extends DbAdapter implements ListInterface,InsertInterface,EditUpdateI
             $this->db["username"] = $this->dbCipher->encrypt($this->requested->username);
             $this->db["email"]    = $this->dbCipher->encrypt($this->requested->email);
             $this->db["passwd"]   = (new CipherGeneric(new PasswordHash()))->hash( $this->requested->passwd );
-            $this->db->table( R::tables('users') )->where("id", $this->requested->id)->update();
+            $this->db->table( R::tables('users') )->where("_id", $this->requested->id)->update();
             $this->db->commit();
         }catch(\Exception $e){
             $this->db->rollBack();